requirements.txt
allure-pytest~=2.13.5
allure-python-commons~=2.13.5
flake8~=7.1.1
pytest~=8.3.3
requests~=2.32.3
selenium~=4.26.1
webdriver-manager~=4.0.2